have separate outboard traits for sync and async
this will allow having an outboard remotely
